# docs/swagger.yaml
definitions:
  entity.Canvassing:
    properties:
      cooldown:
        type: integer
      expire:
        type: integer
      interval:
        type: integer
      is_active:
        type: boolean
      last_auto_blast:
        type: string
      merchant_id:
        type: integer
      radius:
        type: number
    type: object
  entity.CanvassingCall:
    properties:
      created_at:
        type: string
      customer_id:
        type: integer
      deleted_at:
        type: string
      foodtruck_id:
        type: integer
      id:
        type: integer
      notif_id:
        type: integer
      queue_no:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  entity.FoodtruckLocation:
    properties:
      foodtruck_id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  entity.Users:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      platNo:
        type: string
      profilePicture:
        type: string
      updatedAt:
        type: string
      userName:
        type: string
    type: object
...
  entity.UsersLocation:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      user_id:
        type: integer
    type: object
...
  model.RespCallsByFoodtruckID:
    properties:
      created_at:
        type: string
      customer_id:
        type: integer
      deleted_at:
        type: string
      foodtruck_id:
        type: integer
      id:
        type: integer
      latitude_foodtruck:
        type: number
      latitude_user:
        type: number
      longitude_foodtruck:
        type: number
      longitude_user:
        type: number
      name:
        type: string
      notif_id:
        type: integer
      phone:
        type: string
      profile_picture:
        type: string
      queue_no:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.RespCallsByUserID:
    properties:
      created_at:
        type: string
      customer_id:
        type: integer
      deleted_at:
        type: string
      foodtruck_id:
        type: integer
      id:
        type: integer
      ig_account:
        type: string
      latitude_foodtruck:
        type: number
      latitude_user:
        type: number
      logo:
        type: string
      longitude_foodtruck:
        type: number
      longitude_user:
        type: number
      name:
        type: string
      notif_id:
        type: integer
      phone:
        type: string
      plat_no:
        type: string
      queue_no:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.RespCanvassing:
    properties:
      cooldown:
        type: integer
      expire:
        type: integer
      interval:
        type: integer
      is_active:
        type: boolean
      is_auto_blast:
        type: boolean
      last_auto_blast:
        type: string
      last_blast:
        type: string
      merchant_id:
        type: integer
      radius:
        type: number
    type: object
  model.RespNotifByUserID:
    properties:
      created_at:
        type: string
      customer_id:
        type: integer
      foodtruck_id:
        type: integer
      id:
        type: integer
      ig_account:
        type: string
      latitude_foodtruck:
        type: number
      latitude_user:
        type: number
      logo:
        type: string
      longitude_foodtruck:
        type: number
      longitude_user:
        type: number
      name:
        type: string
      plat_no:
        type: string
      status:
        type: string
    type: object
  model.ResponseErrors:
    properties:
      error:
        $ref: '#/definitions/model.errorResponse'
    type: object
  model.ResponseSuccess:
    properties:
      data:
        type: object
    type: object
...
  model.errorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /canvassing:
    get:
      operationId: GetCanvas
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.Canvassing'
        "500":
          description: Inquiry Canvassing Rule Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get Admin's Canvassing Rule (permission = admin)
      tags:
      - Canvassing
    post:
      operationId: CreateCanvass
      parameters:
      - description: merchant_id and last_auto_blast fields are not mandatory
        in: body
        name: canvassing
        required: true
        schema:
          $ref: '#/definitions/entity.Canvassing'
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.Canvassing'
        "422":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Create Canvassing Rule Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Create Canvassing Rule (permission = admin)
      tags:
      - Canvassing
    put:
      operationId: UpdateCanvass
      parameters:
      - description: merchant_id and last_auto_blast fields are not mandatory
        in: body
        name: canvassing
        required: true
        schema:
          $ref: '#/definitions/entity.Canvassing'
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.Canvassing'
        "422":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Update Canvassing Rule Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update Foodtruck's Canvassing Rule (permission = admin)
      tags:
      - Canvassing
  /canvassing/blast:
    post:
      operationId: Blast
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "422":
          description: No users nearby
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Blast Canvassing Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Blast Notification to Users Nearby Manually (permission = merchant)
      tags:
      - Canvassing
  /canvassing/call-status/:call-id/:status:
    put:
      operationId: UpdateStatusCall
      parameters:
      - description: "1"
        in: path
        name: call-id
        required: true
        type: string
      - description: onprocess/accept/reject/expire
        in: path
        name: status
        required: true
        type: string
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.CanvassingCall'
        "500":
          description: Call Status Updated Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update Status Call From Consumer (permission = merchant)
      tags:
      - Canvassing
  /canvassing/call/:call-id/:status:
    put:
      operationId: AnswerCall
      parameters:
      - description: "1"
        in: path
        name: call-id
        required: true
        type: string
      - description: onprocess/accept/reject/expire
        in: path
        name: status
        required: true
        type: string
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.CanvassingCall'
        "500":
          description: Answer Consumer Call Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Answer Call From Consumer (permission = merchant)
      tags:
      - Canvassing
  /canvassing/call/:notif-id:
    post:
      operationId: CallFoodtruck
      parameters:
      - description: "1"
        in: path
        name: notif-id
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: Calling Foodtruck Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Call Foodtruck (permission = consumer)
      tags:
      - Canvassing
  /canvassing/calls:
    get:
      operationId: GetCallsByCustomerID
      parameters:
      - description: accept/request/history
        in: query
        name: status
        required: true
        type: string
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            items:
              $ref: '#/definitions/model.RespCallsByUserID'
            type: array
        "500":
          description: Inquiry Calls Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Inquiry Customer's Calls (permission = consumer)
      tags:
      - Canvassing
  /canvassing/finish/:call-id:
    put:
      operationId: FinishCall
      parameters:
      - description: "1"
        in: path
        name: call-id
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: Finish Consumer Call Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Finish Call From Consumer (permission = merchant)
      tags:
      - Canvassing
  /canvassing/foodtruck:
    get:
      operationId: GetFoodtruckCanvas
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/model.RespCanvassing'
        "500":
          description: Inquiry Canvassing Rule Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get Foodtruck's Canvassing Rule (permission = merchant)
      tags:
      - Canvassing
  /canvassing/foodtruck/calls:
    get:
      operationId: GetCallsByFoodtruckID
      parameters:
      - description: onprocess/accept/request/history
        in: query
        name: status
        required: true
        type: string
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            items:
              $ref: '#/definitions/model.RespCallsByFoodtruckID'
            type: array
        "500":
          description: Inquiry Calls Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Inquiry Foodtruck's Calls (permission = merchant)
      tags:
      - Canvassing
  /canvassing/foodtruck/location:
    put:
      operationId: UpdateFoodtruckLocation
      parameters:
      - description: id fields are not mandatory
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/entity.FoodtruckLocation'
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "422":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Update Foodtruck Location Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update Foodtruck's Location (permission = merchant)
      tags:
      - Canvassing
  /canvassing/foodtruck/location/:id:
    get:
      operationId: GetFoodtruckLocation
      parameters:
      - description: "1"
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.FoodtruckLocation'
        "500":
          description: Inquiry Foodtruck Location Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get Foodtruck's Location (permission = all)
      tags:
      - Canvassing
  /canvassing/notifications:
    get:
      operationId: GetNotifications
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            items:
              $ref: '#/definitions/model.RespNotifByUserID'
            type: array
        "500":
          description: Inquiry Notifications Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Inquiry Users' Notifications (permission = consumer)
      tags:
      - Canvassing
  /canvassing/toggle/auto:
    put:
      operationId: ToggleAutoBlast
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
      security:
      - Token: []
      summary: Toggle Foodtruck Auto Blast (permission = merchant)
      tags:
      - Canvassing
  /canvassing/users/location:
    put:
      operationId: UpdateUsersLocation
      parameters:
      - description: id fields are not mandatory
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/entity.UsersLocation'
      responses:
        "200":
          description: '{"message": "Success"}'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "422":
          description: Invalid request
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: Update Users Location Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Update Users's Location (permission = consumer)
      tags:
      - Canvassing
  /canvassing/users/location/:
    get:
      operationId: GetUserLocation
      responses:
        "200":
          description: '{ "data": Model }'
          schema:
            $ref: '#/definitions/entity.UsersLocation'
        "500":
          description: Inquiry User's Location Failed
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Get User's Location (permission = consumer)
      tags:
      - Canvassing
...
  /merchant/registration-token/{token}:
    post:
      operationId: RegistrationToken
      parameters:
      - description: fcm token from client SDK
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: 'data: Registration Token Successfully'
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
      security:
      - Token: []
      summary: Send Registration Token FCM, Its should be done to archive push notification
        (permission = merchant)
      tags:
      - Merchant
...
  /user/info/{id}:
    get:
      operationId: GetUserByID
      parameters:
      - description: userID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 'data: entity.Users'
          schema:
            $ref: '#/definitions/entity.Users'
      security:
      - Token: []
      summary: Get user profile by ID (permission = superadmin)
      tags:
      - Master User
...
swagger: "2.0"